Finished styling payment form
Used a grid layout for the payment form.

Also had to refactor the label elements in html, again, in order to get the text along with the input forms to align correctly.

I know I should use implicit labels for accessibility reasons, I will look into it later, but this works for now.